BAMPFA-549: collectionobject measuredPart is now free text and dimensionSummary is now computed
